CORE-13926 remove cache evicter
caches register eviction listeners themselves
